Insertion of the template imaNsIdCnt in the supported IMA templates
Now is possible to create Entry with lines on this template
Is inserted the logic to support the verification. is possible to replicate
the PCRs extension done by childs. Is possible having just the list of the host
to verificate its integrity. The solution refers to the Implementation of the IMA
namespace and extension of the measure to parents.