Implementation of APIs for 2 layer attestation
Implementation of a client-verifier that is able to communicate to the cloud verifier
and ask for the verification of the whole host list andgive back
a list regarding the namespace under attestation.
The client verifier will check the equailty between the list executed and golden values.
The client verifer fiels for configuration are inserted. The configuration is in teh mapping.json
file and then dumped in clientverifier.js, the configuraitons are inserted in the keylime.conf