Modification of tha APIs path
Reference namespaces by container id and perform polling on all the
agents because the position of the contianer is not know at priori.
Now the API to contact the agent will require the containerid
and will return the identifier of the ima namespace and the list
of measurements of the namespace.